requirements.txt
boto3
opencv-python
python-dotenv